First implementation of the bootloader update panel for esp32s2.
Added ESP32 boards that have a release in the tinyuf2 repository, had to convert the name of some of them (adafruit_magtag_2.9_grayscale on cpy.org, adafruit_magtag_29gray in the uf2).
The text is incomplete (I'd like to include link to esptool instructions to begin with) and ambiguous (only adafruit boards support the doble-click, so both instructions are given).